requirements.txt
Django==1.9.0
mysql-python
html5lib>=1.0b10
xhtml2pdf>=0.2b1
httplib2>=0.8
pyPdf2>=1.26
Pillow>=2.4.0
reportlab>=3
django-easy-pdf
django-import-export
django-datatables-view
pandas
djangorestframework==3.6.3